requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.30
httpx==0.27.0
python-dotenv==1.0.1
openai==1.25.0
pydantic==2.7.1 